+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## master (unreleased)
 
+### Bugs fixed
+
+* `orchard.inspect`: don't render `Datafy` sections identical to the data they refer to, for nil-valued maps.
+
 ## 0.17.0 (2023-10-24)
 
 ### New features